+++ vojuck.victoria/P4/isUpperTriangular.cpp
@@ -1,6 +1,6 @@
 #include "isUpperTriangular.hpp"
 
-bool isUpperTriangular(int* matrix, size_t rows, size_t cols)
+bool isUpperTriangular(long long * matrix, size_t rows, size_t cols)
 {
   for (size_t i = 0; i < rows; i++)
   {
